backend_pharmacie/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==3.2.0
click==8.3.1
Deprecated==1.3.1
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.135.1
greenlet==3.3.2
h11==0.16.0
idna==3.11
limits==5.8.0
numpy==2.4.3
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
passlib>=1.7.4
psycopg2-binary==2.9.11
requests==2.32.3
redis==5.2.1
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
six==1.17.0
slowapi==0.1.9
SQLAlchemy==2.0.48
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
wrapt==2.1.2

# Code Quality & Testing
black==24.1.1
pylint==3.0.3
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.2
